+++ R/RLearner_classif_fdausc.np.R
@@ -32,13 +32,11 @@ trainLearner.classif.fdausc.np = function(.learner, .task, .subset, .weights = N
   d = getTaskData(.task, subset = .subset, target.extra = TRUE, functionals.as = "matrix")
   fd = getFunctionalFeatures(d$data)
   # transform the data into fda.usc:fdata class type.
-  data.fdclass = fda.usc::fdata(mdata = as.matrix(fd))
+  data.fdclass = fdata(mdata = as.matrix(fd))
 
   par.cv = learnerArgsToControl(list, trim, draw)
-  mod = fda.usc::classif.np(group = d$target, fdataobj = data.fdclass, par.CV = par.cv,
+  mod = classif.np(group = d$target, fdataobj = data.fdclass, par.CV = par.cv,
     par.S = list(w = .weights), ...)
-  # Fix a bug in the package
-  mod$C[[1]] = quote(classif.np)
   return(mod)
 }
 
@@ -46,7 +44,7 @@ trainLearner.classif.fdausc.np = function(.learner, .task, .subset, .weights = N
 predictLearner.classif.fdausc.np = function(.learner, .model, .newdata, ...) {
   # transform the data into fda.usc:fdata class type.
   fd = getFunctionalFeatures(.newdata)
-  nd = fda.usc::fdata(mdata = as.matrix(fd))
+  nd = fdata(mdata = as.matrix(fd))
 
   # predict according to predict.type
   type = ifelse(.learner$predict.type == "prob", "probs", "class")