+++ tests/testthat/test_classif_fdausc.glm.R
@@ -1,7 +1,8 @@
 context("RLearner_classif_fdausc.glm")
 
 test_that("classif_fdausc.glm behaves like original api", {
-  requirePackagesOrSkip("fda.usc", default.method = "load")
+
+  requirePackagesOrSkip("fda.usc", default.method = "attach")
 
   data(phoneme, package = "fda.usc")
   # Use only 10 obs. for 5 classes, as knn training is really slow
@@ -15,10 +16,8 @@ test_that("classif_fdausc.glm behaves like original api", {
   dataf = data.frame(glearn)
   dat = list("df" = dataf, "x" = mlearn)
   # glm sometimes does not converge, we dont want to see that
-  a1 = suppressWarnings(fda.usc::classif.glm(glearn ~ x, data = dat))
+  a1 = suppressWarnings(classif.glm(glearn ~ x, data = dat))
 
-  # FIXME: code looks strange here? the quote?
-  a1$C[[1]] = quote(classif.glm)
   p1 = predict(a1, list("x" = mtest))
   p2 = predict(a1, list("x" = mlearn))
 
@@ -32,7 +31,6 @@ test_that("classif_fdausc.glm behaves like original api", {
   ftest = makeFunctionalData(phtst, fd.features = NULL, exclude.cols = "label")
   task = makeClassifTask(data = fdata, target = "label")
 
-  set.seed(getOption("mlr.debug.seed"))
   # glm sometimes does not converge, we dont want to see that
   m = suppressWarnings(train(lrn, task))
   cp = predict(m, newdata = ftest)
@@ -40,6 +38,7 @@ test_that("classif_fdausc.glm behaves like original api", {
 
   cp2 = predict(m, newdata = fdata)
   cp2 = unlist(cp2$data$response, use.names = FALSE)
+
   # check if the output from the original API matches the mlr learner's output
   expect_equal(as.character(cp2), as.character(p2))
   expect_equal(as.character(cp), as.character(p1))